new file for open file table

+++ kern/arch/mips/syscall/syscall.c
@@ -113,7 +113,7 @@ syscall(struct trapframe *tf)
 	    /* Add stuff here */
 		case SYS_open:
 		err = open((char *)tf->tf_a0,
-				(int)tf->tf_a1,
+				tf->tf_a1,
                 (mode_t)tf->tf_a2);
 		break;
 /*
